app/requirements.txt
line-bot-sdk==3.5.0
boto3==1.34.14
requests==2.31.0